+++ augmentos_core/app/src/main/java/com/augmentos/augmentos_core/smarterglassesmanager/smartglassescommunicators/evenrealities/connection/ConnectionManager.java
@@ -1,8 +1,26 @@
+/**
+ * ConnectionManager is responsible for orchestrating communication between the Even Realities smart glasses
+ * and the application. It manages the left and right BLE (Bluetooth Low Energy) connections, handles
+ * command dispatching, response listening, and provides an optional synchronous interface using futures.
+ *
+ * This class ensures that commands are sent to the correct device (left/right/both), manages a queue
+ * to prevent overlapping commands with conflicting responses, and resolves responses using predefined
+ * headers and response parsers.
+ *
+ * It also supports both asynchronous and synchronous command execution (via `sendCommand` and `sendAndWait`)
+ * and delegates all BLE operations to the underlying `Connection` instances.
+ *
+ * Designed to serve as the main communication bridge for SDK-like integrations with Even Realities G1 (firmware 1.5.0),
+ * with support for future enhancements like heartbeat monitoring, retries, or extended device status.
+ */
+
+package com.augmentos.augmentos_core.smarterglassesmanager.smartglassescommunicators.evenrealities.connection;
+
 import java.util.*;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 
-import com.augmentos.augmentos_core.smarterglassesmanager.smartglassescommunicators.evenrealities.connection.BleConfig;
+import com.augmentos.augmentos_core.smarterglassesmanager.smartglassescommunicators.evenrealities.connection.ConnectionConfig;
 import com.augmentos.augmentos_core.smarterglassesmanager.smartglassescommunicators.evenrealities.connection.Connection;
 import com.augmentos.augmentos_core.smarterglassesmanager.smartglassescommunicators.evenrealities.evenos.EvenOsBase;
 import com.augmentos.augmentos_core.smarterglassesmanager.smartglassescommunicators.evenrealities.evenos.EvenOsCommand;
@@ -27,7 +45,7 @@ public class ConnectionManager {
         uartTxCharUuid = UUID.fromString("6E400002-B5A3-F393-E0A9-E50E24DCCA9E");
         uartRxCharUuid = UUID.fromString("6E400003-B5A3-F393-E0A9-E50E24DCCA9E");
         int mtu = 512;
-        BleConfig config = new BleConfig(UartServiceUuid, uartTxCharUuid, uartRxCharUuid, mtu);
+        ConnectionConfig config = new ConnectionConfig(UartServiceUuid, uartTxCharUuid, uartRxCharUuid, mtu);
 
         this.leftConnection = new Connection(context, smartGlassesDevice.getLeftDevice(), config);
         this.rightConnection = new Connection(context, smartGlassesDevice.getRightDevice(), config);
