+++ augmentos_core/app/src/main/java/com/augmentos/augmentos_core/smarterglassesmanager/smartglassescommunicators/evenrealities/exception/BleInitializationException.java
@@ -1,0 +1,12 @@
+package com.augmentos.augmentos_core.smarterglassesmanager.smartglassescommunicators.evenrealities.exception;
+
+
+/**
+ * Exception thrown when the BLE initialization fails.
+ * Wrong UUIDs, wrong device, etc.
+ */
+public class BleInitializationException extends RuntimeException {
+    public BleInitializationException(String message) {
+        super(message);
+    }
+}
